BML: Alters BML_CLIENT:_CONFIG add time_life_delay_minutes
The upcoming revisions are dependent on the addition
to the BML definition of sClientConfig structure.

Adds time_life_delay_minutes to the sClientConfig
structure like previously described (in bml_defs).

+++ controller/src/beerocks/bml/bml_defs.h
@@ -593,10 +593,11 @@ struct BML_CLIENT_CONFIG {
     int8_t selected_bands;
 
     // Optional parameter,
-    // Determines the period of time after which the client configuration should be cleared,
+    // Determines the period of time after which the client configuration
     //   0 - Never age.
-    //  -1 - Not Configured (optional).
-    int32_t time_life_delay_minutes;
+    //  -1 - Not Configured.
+    //  >0 - Hardcoded timelife
+    int8_t time_life_delay_minutes;
 };
 
 struct BML_CLIENT {
